src/azure-container-apps/app/requirements.txt
Flask
requests
gunicorn
azure-cosmos
python-dotenv
